requirements.txt
absl-py==0.1.10
aiohttp==1.0.5
appdirs==1.4.3
astroid==2.3.3
async-timeout==2.0.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
bs4==0.0.1
cffi==1.9.1
chardet==3.0.4
colorama==0.3.7
cryptography==1.7.2
cycler==0.10.0
decorator==4.0.11
discord.py==0.16.12
distlib==0.3.3
entrypoints==0.2.2
enum34==1.1.6
ez-setup==0.9
filelock==3.2.1
future==0.18.2
html5lib==0.9999999
HTMLParser==0.0.2
idna==2.2
importlib-metadata==2.1.1
importlib-resources==3.2.1
inflection==0.3.1
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.3.21
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
Keras==2.1.4
lazy-object-proxy==1.4.3
List==1.3.0
lxml==3.7.3
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.0.0
mccabe==0.6.1
mistune==0.7.3
more-itertools==2.5.0
multidict==4.5.2
nbconvert==5.1.1
nbformat==4.2.0
ndg-httpsclient==0.4.2
nltk==3.2.4
notebook==4.3.1
numpy==1.14.0+mkl
oauthlib==3.1.0
packaging==16.8
pandas==0.19.2
pandas-datareader==0.3.0.post0
pandocfilters==1.4.1
pickleshare==0.7.4
platformdirs==2.0.2
prompt-toolkit==1.0.9
protobuf==3.5.1
pyasn1==0.1.9
pycparser==2.17
Pygments==2.1.3
pylint==2.4.4
pymongo==3.4.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
PyTumblr==0.1.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
Quandl==3.0.1
requests==2.13.0
requests-file==1.4.1
requests-ftp==0.3.1
requests-oauthlib==1.3.0
scikit-learn==0.18.1
scipy==1.0.0
seaborn==0.7.1
selenium==3.9.0
simplegeneric==0.8.1
six==1.13.0
sklearn==0.0
tensorflow==1.4.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
testpath==0.3
textblob==0.12.0
tornado==4.4.2
traitlets==4.3.1
typed-ast==1.4.0
urllib3==1.23
virtualenv==20.9.0
wcwidth==0.1.7
websockets==3.4
Werkzeug==0.14.1
widgetsnbextension==1.2.6
win-unicode-console==0.5
wrapt==1.11.2
zipp==1.2.0
